Re-org for test definitions
